% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactslider.R
\name{reactsliderInput}
\alias{reactsliderInput}
\title{React Slider}
\usage{
reactsliderInput(
  inputId,
  data = NULL,
  min = NULL,
  max = NULL,
  step = 1,
  start = NULL,
  tooltip = c("always", "onClick", "never"),
  prefix = "",
  postfix = "",
  labels = list(),
  sticky = FALSE,
  orientation = c("horizontal", "vertical"),
  chart_tooltip = FALSE,
  width = 400,
  default = 0
)
}
\description{
Create a react slider
}
